backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Data Validation (Pydantic v2 - works perfectly with Python 3.12!)
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.1.1
typing-extensions>=4.12.0

# HTTP Client & File Uploads
httpx==0.28.1
python-multipart==0.0.20

# AI Library
google-generativeai==0.8.3

# ML Libraries (now with pre-built wheels for Python 3.12!)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Messaging (optional)
twilio==9.4.0

# Text-to-Speech (optional)
gtts==2.5.4

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# NOTES:
# - Python 3.12 has pre-built wheels for ALL packages!
# - Pydantic v2 works perfectly (no Rust compiler needed)
# - All ML features now fully functional with numpy/scikit-learn
# - No compilation required - everything installs smoothly
